+++ lib/overcommit/hook/pre_commit/go_lint.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `golint` against any modified Golang files.
   class GoLint < Base
     def run
-      result = execute([executable] + applicable_files)
+      result = execute(command + applicable_files)
       # Unfortunately the exit code is always 0
       return :pass if result.stdout.empty?