+++ lib/overcommit/hook/pre_commit/js_hint.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `jshint` against any modified JavaScript files.
   class JsHint < Base
     def run
-      result = execute([executable] + applicable_files)
+      result = execute(command + applicable_files)
       output = result.stdout
 
       return :pass if output.empty?