+++ lib/overcommit/hook/pre_commit/jsx_hint.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `jsxhint` against any modified JSX files.
   class JsxHint < Base
     def run
-      result = execute([executable] + applicable_files)
+      result = execute(command + applicable_files)
       output = result.stdout
 
       return :pass if output.empty?