+++ lib/overcommit/hook/pre_commit/python_flake8.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `flake8` against any modified Python files.
   class PythonFlake8 < Base
     def run
-      result = execute([executable] + applicable_files)
+      result = execute(command + applicable_files)
       return :pass if result.success?
 
       [:fail, result.stdout]